+++ Fotogroep Waalre/View/PhotoClubsView/PhotoClubListView.swift
@@ -59,7 +59,7 @@ struct PhotoClubListView: View {
                 let andersBackgroundContext = PersistenceController.shared.container.newBackgroundContext()
                 andersBackgroundContext.name = "Anders refresh"
                 andersBackgroundContext.mergePolicy = NSMergePolicy.mergeByPropertyObjectTrump
-                _ = AndersMembersProvider(bgContext: andersBackgroundContext)
+                // _ = AndersMembersProvider(bgContext: andersBackgroundContext) TODO re-enable here and add at startup
 
                 // load all current/former members of Fotogroep Waalre
                 let fgwBackgroundContext = PersistenceController.shared.container.newBackgroundContext()